requirements.txt
alembic==1.4.2
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
dominate==2.5.1
Flask==1.1.2
flask-blueprint==1.3.0
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
Naked==0.1.31
Pillow==7.1.2
pycryptodome==3.9.7
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
requests==2.23.0
shellescape==3.8.1
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
